+++ dcmjpeg/apps/dcmcjpeg.cc
@@ -22,9 +22,9 @@
  *  Purpose: Compress DICOM file
  *
  *  Last Update:      $Author: meichel $
- *  Update Date:      $Date: 2001-11-19 15:13:22 $
+ *  Update Date:      $Date: 2001-12-20 10:41:45 $
  *  Source File:      $Source: /export/gitmirror/dcmtk-git/../dcmtk-cvs/dcmtk/dcmjpeg/apps/dcmcjpeg.cc,v $
- *  CVS/RCS Revision: $Revision: 1.2 $
+ *  CVS/RCS Revision: $Revision: 1.3 $
  *  Status:           $State: Exp $
  *
  *  CVS/RCS Log at end of file
@@ -529,7 +529,7 @@ int main(int argc, char *argv[])
       opt_uidcreation,
       opt_verbose,
       opt_huffmanOptimize,
-      opt_smoothing,   
+      (int) opt_smoothing,   
       opt_compressedBits,   
       (Uint32) opt_fragmentSize,
       opt_createOffsetTable,
@@ -706,7 +706,10 @@ int main(int argc, char *argv[])
 /*
  * CVS/RCS Log:
  * $Log: dcmcjpeg.cc,v $
- * Revision 1.2  2001-11-19 15:13:22  meichel
+ * Revision 1.3  2001-12-20 10:41:45  meichel
+ * Fixed warnings reported by Sun CC 2.0.1
+ *
+ * Revision 1.2  2001/11/19 15:13:22  meichel
  * Introduced verbose mode in module dcmjpeg. If enabled, warning
  *   messages from the IJG library are printed on ofConsole, otherwise
  *   the library remains quiet.
